Revert "fix(collect-review-texts): rename file when criticName override changes slug (#290)"
This reverts commit cfa76f968a27b11dc2e2d27a8efa7c9145723ce3.
