+++ .github/workflows/test.yml
@@ -335,7 +335,7 @@ jobs:
           token: ${{ secrets.REVIEW_TEXTS_TOKEN }}
       - name: Run unit tests (no-data-dependency)
         run: |
-          node --test tests/unit/regression-guards.test.mjs tests/unit/sample.test.mjs tests/unit/sanity.test.mjs tests/unit/score-conversion-rules.test.mjs tests/unit/score-routing.test.mjs tests/unit/email-capture-integrity.test.mjs tests/unit/placeholder-images.test.mjs tests/unit/compute-critic-score.test.mjs tests/unit/wrong-production-autoclear.test.mjs tests/unit/clear-failure-flags.test.mjs tests/unit/is-includable-for-rebuild.test.mjs tests/unit/verify-fetched-url.test.mjs tests/unit/review-write-guard.test.mjs tests/unit/theater-scorecard-validators.test.mjs tests/unit/theater-scorecard-regression.test.mjs tests/unit/bww-roundup-extractor.test.mjs tests/unit/bww-rr-discover.test.mjs tests/unit/protected-fields-sync.test.mjs tests/unit/protected-fields-three-way-sync.test.mjs tests/unit/human-review-score-provisional.test.mjs tests/unit/ingest-manual-review-fields.test.mjs tests/unit/ingest-manual-review-collision.test.mjs tests/unit/recover-manual-review-score.test.mjs tests/unit/poller-blocklist.test.mjs tests/unit/block-review-cli.test.mjs tests/unit/no-reset-rsync-push-pattern.test.mjs tests/unit/scraper-cookie-wiring.test.mjs tests/unit/opening-night-poll-cadence.test.mjs tests/unit/opening-night-bd-gate.test.mjs tests/unit/opening-night-serp-backup-integration.test.mjs tests/unit/opening-night-checks-juan-a-ramirez-bypass.test.mjs tests/unit/opening-night-bypass-corpus.test.mjs tests/unit/orchestrator-guardian-dispatch.test.mjs tests/unit/pending-strand-routing.test.mjs tests/unit/review-text-scoreable.test.mjs tests/unit/paywall-detector.test.mjs tests/unit/content-quality-paywall-subscriber.test.mjs tests/unit/content-quality-wrong-production-autocleared.test.mjs tests/unit/pending-gap-classification.test.mjs tests/unit/broadcast-state.test.mjs tests/unit/reconcile-broadcast-state.test.mjs tests/unit/orphan-pending-decision.test.mjs tests/unit/input-builder-market-label.test.mjs tests/unit/multi-critic-serp.test.mjs tests/unit/llm-fallback-partial.test.mjs tests/unit/nypost-inline-star-recovery.test.mjs tests/unit/tour-forward-tense.test.mjs tests/unit/gather-reviews-collision.test.mjs tests/unit/venue-aliases.test.mjs tests/unit/long-runner-registry.test.mjs tests/unit/url-year-conflict-hint.test.mjs tests/unit/market-routing.test.mjs tests/unit/playbill-verdict-discover.test.mjs tests/unit/lbo-roundup-page-validation.test.mjs tests/unit/lbo-individual-star-extraction.test.mjs tests/unit/is-likely-stale-roundup-flag.test.mjs tests/unit/is-likely-stale-suspected-misattribution.test.mjs tests/unit/is-likely-stale-wrong-show.test.mjs tests/unit/is-likely-stale-wrong-production.test.mjs tests/unit/wrong-production-setters-honor-manual-clear.test.mjs tests/unit/is-scoreable-suspected-misattribution.test.mjs tests/unit/seo-anomaly-detection.test.mjs tests/unit/poller-idempotency.test.mjs tests/unit/rename-review-file-to-match-critic.test.mjs tests/unit/article-extractor-multi-article.test.mjs
+          node --test tests/unit/regression-guards.test.mjs tests/unit/sample.test.mjs tests/unit/sanity.test.mjs tests/unit/score-conversion-rules.test.mjs tests/unit/score-routing.test.mjs tests/unit/email-capture-integrity.test.mjs tests/unit/placeholder-images.test.mjs tests/unit/compute-critic-score.test.mjs tests/unit/wrong-production-autoclear.test.mjs tests/unit/clear-failure-flags.test.mjs tests/unit/is-includable-for-rebuild.test.mjs tests/unit/verify-fetched-url.test.mjs tests/unit/review-write-guard.test.mjs tests/unit/theater-scorecard-validators.test.mjs tests/unit/theater-scorecard-regression.test.mjs tests/unit/bww-roundup-extractor.test.mjs tests/unit/bww-rr-discover.test.mjs tests/unit/protected-fields-sync.test.mjs tests/unit/protected-fields-three-way-sync.test.mjs tests/unit/human-review-score-provisional.test.mjs tests/unit/ingest-manual-review-fields.test.mjs tests/unit/ingest-manual-review-collision.test.mjs tests/unit/recover-manual-review-score.test.mjs tests/unit/poller-blocklist.test.mjs tests/unit/block-review-cli.test.mjs tests/unit/no-reset-rsync-push-pattern.test.mjs tests/unit/scraper-cookie-wiring.test.mjs tests/unit/opening-night-poll-cadence.test.mjs tests/unit/opening-night-bd-gate.test.mjs tests/unit/opening-night-serp-backup-integration.test.mjs tests/unit/opening-night-checks-juan-a-ramirez-bypass.test.mjs tests/unit/opening-night-bypass-corpus.test.mjs tests/unit/orchestrator-guardian-dispatch.test.mjs tests/unit/pending-strand-routing.test.mjs tests/unit/review-text-scoreable.test.mjs tests/unit/paywall-detector.test.mjs tests/unit/content-quality-paywall-subscriber.test.mjs tests/unit/content-quality-wrong-production-autocleared.test.mjs tests/unit/pending-gap-classification.test.mjs tests/unit/broadcast-state.test.mjs tests/unit/reconcile-broadcast-state.test.mjs tests/unit/orphan-pending-decision.test.mjs tests/unit/input-builder-market-label.test.mjs tests/unit/multi-critic-serp.test.mjs tests/unit/llm-fallback-partial.test.mjs tests/unit/nypost-inline-star-recovery.test.mjs tests/unit/tour-forward-tense.test.mjs tests/unit/gather-reviews-collision.test.mjs tests/unit/venue-aliases.test.mjs tests/unit/long-runner-registry.test.mjs tests/unit/url-year-conflict-hint.test.mjs tests/unit/market-routing.test.mjs tests/unit/playbill-verdict-discover.test.mjs tests/unit/lbo-roundup-page-validation.test.mjs tests/unit/lbo-individual-star-extraction.test.mjs tests/unit/is-likely-stale-roundup-flag.test.mjs tests/unit/is-likely-stale-suspected-misattribution.test.mjs tests/unit/is-likely-stale-wrong-show.test.mjs tests/unit/is-likely-stale-wrong-production.test.mjs tests/unit/wrong-production-setters-honor-manual-clear.test.mjs tests/unit/is-scoreable-suspected-misattribution.test.mjs tests/unit/seo-anomaly-detection.test.mjs tests/unit/poller-idempotency.test.mjs tests/unit/article-extractor-multi-article.test.mjs
 
           npx tsx --test tests/unit/engine.test.ts tests/unit/scoring-parity.test.ts tests/unit/llm-scoring-stale-roundup-flag.test.mjs
